feat(toolkit-rust): Hard code http scheme for the tool URL (#78)
Using `https` as the tool scheme will not work in internal services
running in Kubernetes without setting up a TLS server with signed
certificates.

This PR removes the hard coded scheme and instead use the scheme and
host available in the HTTP headers `X-Forwarded-Host` and
`X-Forwarded-Proto`. If those aren't available, it defaults to the
previously used values.

The `http` schema support will be polished up when we implement #61.

Closes #77

---------

Co-authored-by: Christos KK Loverdos <loverdos@gmail.com>